01-hello-world/pyproject.toml
[project]
name = "genai-hello-world"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "sap-ai-sdk-gen==5.8.0",
    "python-dotenv",
]

[tool.uv]
# uv will use this pyproject to create an isolated environment
